pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
    "flask>=3.1.0",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2024.2",
    "requests>=2.32.3",
]
